Encrypted data bags instead of chef-vault
Chef vault is cool, but imposes a change in the flow. To make it work
you need to:

1. Bootstrap and add server to a role
2. Enrypt your secret sith all the keys of the role
3. Run your recipe where you use your secret

With encrypted data bags you do all the above in one step i.e.
bootstraping takes care of copying the secret_key, adding the node into
a role and running the recipe where you use it